+++ src/components/TabNav/example/index.js
@@ -34,26 +34,14 @@ export default [
   {
     isPadded: true,
     examples: [
-      <TabNav activeTab={ 'first' } style={ ['small'] }>
-        <TabNav.Tab
-          tabId='first'
-          onClick={ function() {
-            alert('clicked first');  //eslint-disable-line
-          } }>
+      <TabNav activeTab={ 'second' } style={ ['small'] }>
+        <TabNav.Tab>
           Tab #1
         </TabNav.Tab>
-        <TabNav.Tab
-          tabId='second'
-          onClick={ function() {
-            alert('clicked second');  //eslint-disable-line
-          } }>
+        <TabNav.Tab>
           Tab #2
         </TabNav.Tab>
-        <TabNav.Tab
-          tabId='third'
-          onClick={ function() {
-            alert('clicked third');  //eslint-disable-line
-          } }>
+        <TabNav.Tab>
           Tab #3
         </TabNav.Tab>
       </TabNav>,
@@ -63,25 +51,13 @@ export default [
     isPadded: true,
     examples: [
       <TabNav activeTab={ 'first' } style={ ['small', 'center'] }>
-        <TabNav.Tab
-          tabId='first'
-          onClick={ function() {
-            alert('clicked first');  //eslint-disable-line
-          } }>
+        <TabNav.Tab>
           Tab #1
         </TabNav.Tab>
-        <TabNav.Tab
-          tabId='second'
-          onClick={ function() {
-            alert('clicked second');  //eslint-disable-line
-          } }>
+        <TabNav.Tab>
           Tab #2
         </TabNav.Tab>
-        <TabNav.Tab
-          tabId='third'
-          onClick={ function() {
-            alert('clicked third');  //eslint-disable-line
-          } }>
+        <TabNav.Tab>
           Tab #3
         </TabNav.Tab>
       </TabNav>,
@@ -91,25 +67,13 @@ export default [
     isPadded: true,
     examples: [
       <TabNav activeTab={ 'first' } style={ ['small', 'sub'] }>
-        <TabNav.Tab
-          tabId='first'
-          onClick={ function() {
-            alert('clicked first');  //eslint-disable-line
-          } }>
+        <TabNav.Tab>
           Tab #1
         </TabNav.Tab>
-        <TabNav.Tab
-          tabId='second'
-          onClick={ function() {
-            alert('clicked second');  //eslint-disable-line
-          } }>
+        <TabNav.Tab>
           Tab #2
         </TabNav.Tab>
-        <TabNav.Tab
-          tabId='third'
-          onClick={ function() {
-            alert('clicked third');  //eslint-disable-line
-          } }>
+        <TabNav.Tab>
           Tab #3
         </TabNav.Tab>
       </TabNav>,