+++ src/components/Table/example/index.js
@@ -7,87 +7,144 @@ export default [
   {
     isPadded: true,
     examples: [
-      <Table
-        data={
-          [
-            ['Header CTA', '12%', 'Paused'],
-            ['Shorter Contact Form', '4%', 'Draft'],
-            ['Larger search bar', '6.7%', 'Paused'],
-            ['Center aligned headline', '9.3%', 'Running'],
-          ]
-        }
-        headings={
-          [
-            'Experiment',
-            'Conversion Rate',
-            'Status',
-          ]
-        }>
+      <Table>
+        <Table.THead>
+          <Table.TR>
+            <Table.TH>Experiment</Table.TH>
+            <Table.TH>Conversion Rate</Table.TH>
+            <Table.TH>Status</Table.TH>
+          </Table.TR>
+        </Table.THead>
+        <Table.TBody>
+          <Table.TR>
+            <Table.TD>Header CTA</Table.TD>
+            <Table.TD width="20%">12%</Table.TD>
+            <Table.TD>Paused</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Shorter Contact Form</Table.TD>
+            <Table.TD>4%</Table.TD>
+            <Table.TD>Draft</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Larger search bar</Table.TD>
+            <Table.TD>6.7%</Table.TD>
+            <Table.TD>Paused</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Center aligned headline</Table.TD>
+            <Table.TD>9.3%</Table.TD>
+            <Table.TD>Running</Table.TD>
+          </Table.TR>
+        </Table.TBody>
       </Table>,
     ],
   },
   {
     isPadded: true,
     examples: [
-      <Table
-        data={
-          [
-            ['Header CTA', '12%', 'Running'],
-            ['Shorter Contact Form', '4%', 'Draft'],
-            ['Larger search bar', '6.7%', 'Paused'],
-            ['Center aligned headline', '9.3%', 'Running'],
-          ]
-        }
-        headings={
-          [
-            'Experiment',
-            'Conversion Rate',
-            'Status',
-          ]
-        }
-        numberedColumns={ [1] }
-        style="rule">
+      <Table style="rule">
+        <Table.THead>
+          <Table.TR>
+            <Table.TH>Experiment</Table.TH>
+            <Table.TH isNumerical={ true }>Conversion Rate</Table.TH>
+            <Table.TH>Status</Table.TH>
+          </Table.TR>
+        </Table.THead>
+        <Table.TBody>
+          <Table.TR>
+            <Table.TD>Header CTA</Table.TD>
+            <Table.TD isNumerical={ true }>12%</Table.TD>
+            <Table.TD>Paused</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Shorter Contact Form</Table.TD>
+            <Table.TD isNumerical={ true }>4%</Table.TD>
+            <Table.TD>Draft</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Larger search bar</Table.TD>
+            <Table.TD isNumerical={ true }>6.7%</Table.TD>
+            <Table.TD>Paused</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Center aligned headline</Table.TD>
+            <Table.TD isNumerical={ true }>9.3%</Table.TD>
+            <Table.TD>Running</Table.TD>
+          </Table.TR>
+        </Table.TBody>
       </Table>,
     ],
   },
   {
     isPadded: true,
     examples: [
-      <Table
-        data={
-          [
-            ['Header CTA', '12%', <Button size="small">Start</Button>],
-            ['Shorter Contact Form', '4%', <Button size="small">Start</Button>],
-            ['Larger search bar', '6.7%', <Button size="small">Start</Button>],
-            ['Center aligned headline', '9.3%', <Button size="small">Start</Button>],
-          ]
-        }
-        headings={
-          [
-            'Experiment',
-            'Conversion Rate',
-            'Status',
-          ]
-        }
-        numberedColumns={ [1] }
-        collapsedColumns={ [2] }
-        style="wall">
+      <Table style="rule-no-bottom-border">
+        <Table.THead>
+          <Table.TR>
+            <Table.TH>Experiment</Table.TH>
+            <Table.TH isNumerical={ true }>Conversion Rate</Table.TH>
+            <Table.TH>Status</Table.TH>
+          </Table.TR>
+        </Table.THead>
+        <Table.TBody>
+          <Table.TR>
+            <Table.TD>Header CTA</Table.TD>
+            <Table.TD isNumerical={ true }>12%</Table.TD>
+            <Table.TD>Paused</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Shorter Contact Form</Table.TD>
+            <Table.TD isNumerical={ true }>4%</Table.TD>
+            <Table.TD>Draft</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Larger search bar</Table.TD>
+            <Table.TD isNumerical={ true }>6.7%</Table.TD>
+            <Table.TD>Paused</Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD>Center aligned headline</Table.TD>
+            <Table.TD isNumerical={ true }>9.3%</Table.TD>
+            <Table.TD>Running</Table.TD>
+          </Table.TR>
+        </Table.TBody>
       </Table>,
     ],
   },
   {
     isPadded: true,
     examples: [
-      <Table
-        data={
-          [
-            ['Header CTA', '12%', 'Running'],
-            ['Shorter Contact Form', '4%', 'Draft'],
-            ['Larger search bar', '6.7%', 'Paused'],
-            ['Center aligned headline', '9.3%', 'Running'],
-          ]
-        }
-        style="wall">
+      <Table style="wall">
+        <Table.THead>
+          <Table.TR>
+            <Table.TH>Experiment</Table.TH>
+            <Table.TH isNumerical={ true } width="20%">Conversion Rate</Table.TH>
+            <Table.TH isCollapsed={ true }>Status</Table.TH>
+          </Table.TR>
+        </Table.THead>
+        <Table.TBody>
+          <Table.TR>
+            <Table.TD verticalAlign="middle">Header CTA</Table.TD>
+            <Table.TD isNumerical={ true } verticalAlign="middle">12%</Table.TD>
+            <Table.TD verticalAlign="middle"><Button size="small">Start</Button></Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD verticalAlign="middle">Shorter Contact Form</Table.TD>
+            <Table.TD isNumerical={ true } verticalAlign="middle">4%</Table.TD>
+            <Table.TD verticalAlign="middle"><Button size="small">Start</Button></Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD verticalAlign="middle">Larger search bar</Table.TD>
+            <Table.TD isNumerical={ true } verticalAlign="middle">6.7%</Table.TD>
+            <Table.TD verticalAlign="middle"><Button size="small">Start</Button></Table.TD>
+          </Table.TR>
+          <Table.TR>
+            <Table.TD verticalAlign="middle">Center aligned headline</Table.TD>
+            <Table.TD isNumerical={ true } verticalAlign="middle">9.3%</Table.TD>
+            <Table.TD verticalAlign="middle"><Button size="small">Start</Button></Table.TD>
+          </Table.TR>
+        </Table.TBody>
       </Table>,
     ],
   },